Make rewind/fast-forward buttons more discoverable
