+++ packages/replay-next/components/sources/SourceListRowMouseEvents.module.css
@@ -1,0 +1,24 @@
+.PositionLeft,
+.PositionRight {
+  position: absolute;
+  z-index: 5;
+  overflow: visible;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  background-color: var(--body-bgcolor);
+}
+.PositionLeft {
+  font-family: var(--font-family-monospace);
+  font-size: var(--font-size-small);
+  width: calc(var(--line-number-size) + 2ch - 0.25rem);
+  justify-content: flex-end;
+}
+.PositionRight {
+  left: calc(var(--source-line-number-offset) + var(--source-hit-count-offset) - 0.25rem);
+}
+
+.Background {
+  border-radius: 0.25rem;
+  box-shadow: 0 0 0 2px var(--body-bgcolor);
+}
